스코어 창 이벤트
스코어창에서 클릭시 점수 및 경험치 이벤트 빠르게 진행

+++ MioMini01/Assets/Scripts/MiniGameManager.cs
@@ -13,7 +13,7 @@ public class MiniGameManager : MonoBehaviour
     public Sprite[] LotMio;
     public List<Sprite> mioList;
 
-    public float Score = 0;
+    public int Score = 0;
 
     public float fiverScore = 0;
     public bool IsFiver;
